feat: 视频上新

+++ model/sqlite_model.go
@@ -39,7 +39,7 @@ type DemandVideo struct {
 	Size       int64  `column:"size" json:"size"`              // 文件大小
 	Width      int    `column:"width" json:"width"`            // 视频宽度
 	Height     int    `column:"height" json:"height"`          // 视频高度
-	Duration   int    `column:"duration" json:"duration"`      // 视频时间
+	Duration   float64    `column:"duration" json:"duration"`      // 视频时间
 	Describe   string `column:"describe" json:"describe"`      // 描述
 	Title      string `column:"title" json:"title"`            // 视频标题
 	FFmpegJSON string `column:"ffmpeg_json" json:"FFmpegJSON"` // 视频流信息
